requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pymupdf==1.23.8
python-multipart==0.0.6
sentence-transformers==2.2.2
groq==0.4.2
openai==1.3.7
google-generativeai==0.3.2
streamlit==1.28.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
aiofiles==23.2.1 